check if server is up when init index is start

+++ src/main/java/io/github/nooshhub/dao/JdbcDao.java
@@ -88,6 +88,11 @@ public class JdbcDao {
 	 * @param indexName index name
 	 */
 	public void init(String indexName) {
+		if (!this.elasticsearchDao.isServerUp(indexName)) {
+			logger.error("Elasticsearch server is not accessible, please Check.");
+			return;
+		}
+
 		jdbcMetrics();
 
 		this.espipeTimerDao.delete(indexName);
